+++ src/main/java/com/ryan/roomreservationservice/domain/PaymentHistory.java
@@ -4,6 +4,7 @@ import com.ryan.roomreservationservice.domain.enums.PaymentMethod;
 import com.ryan.roomreservationservice.domain.enums.PaymentStatus;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
 @Getter
@@ -14,11 +15,11 @@ public class PaymentHistory {
     private Reservation reservation;
     private PaymentMethod paymentMethod;
     private String transactionId;
-    private Long amount;
+    private BigDecimal amount;
     private String receipt;
     private Instant createdAt;
 
-    public PaymentHistory(PaymentStatus status, Member member, Reservation reservation,  PaymentMethod paymentMethod, String transactionId, Long amount, String receipt) {
+    public PaymentHistory(PaymentStatus status, Member member, Reservation reservation, PaymentMethod paymentMethod, String transactionId, BigDecimal amount, String receipt) {
         this.status = status;
         this.member = member;
         this.reservation = reservation;
