+++ src/main/java/com/ryan/roomreservationservice/domain/PaymentProcess.java
@@ -2,8 +2,10 @@ package com.ryan.roomreservationservice.domain;
 
 import com.ryan.roomreservationservice.domain.enums.PaymentMethod;
 
+import java.math.BigDecimal;
+
 public interface PaymentProcess {
     boolean support(PaymentMethod paymentMethod);
-    PaymentInfo pay(Reservation reservation, long amount);
-    void cancel(String transactionId, long refundAmount);
+    PaymentInfo pay(Reservation reservation, BigDecimal amount);
+    void cancel(String transactionId, BigDecimal refundAmount);
 }
