+++ src/test/java/com/ryan/roomreservationservice/domain/RoomTest.java
@@ -2,6 +2,7 @@ package com.ryan.roomreservationservice.domain;
 
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.ZoneId;
 
@@ -14,7 +15,7 @@ class RoomTest {
         // given(준비): 어떠한 데이터가 준비되었을 때
         ZoneId zoneId = ZoneId.of("Asia/Seoul");
         String name = "그린룸";
-        long price = 300000;
+        BigDecimal price = BigDecimal.valueOf(300000);
 
         Room room = new Room(zoneId, name, price);
 
@@ -22,10 +23,10 @@ class RoomTest {
         LocalDateRange localDateRange = new LocalDateRange(LocalDate.of(2024, 5, 19), LocalDate.of(2024, 5, 20));
 
         // when(실행): 어떠한 함수를 실행하면
-        long refundAmount = room.calculateRoomRefundAmount(localDate, localDateRange);
+        BigDecimal refundAmount = room.calculateRoomRefundAmount(localDate, localDateRange);
 
         // then(검증): 어떠한 결과가 나와야 한다.
-        assertThat(refundAmount).isEqualTo(300000);
+        assertThat(refundAmount).isEqualTo(BigDecimal.valueOf(300000));
     }
 
     @Test
@@ -33,7 +34,7 @@ class RoomTest {
         // given(준비): 어떠한 데이터가 준비되었을 때
         ZoneId zoneId = ZoneId.of("Asia/Seoul");
         String name = "그린룸";
-        long price = 300000;
+        BigDecimal price = BigDecimal.valueOf(300000);
 
         Room room = new Room(zoneId, name, price);
 
@@ -41,9 +42,9 @@ class RoomTest {
         LocalDateRange localDateRange = new LocalDateRange(LocalDate.of(2024, 5, 19), LocalDate.of(2024, 5, 20));
 
         // when(실행): 어떠한 함수를 실행하면
-        long refundAmount = room.calculateRoomRefundAmount(localDate, localDateRange);
+        BigDecimal refundAmount = room.calculateRoomRefundAmount(localDate, localDateRange);
 
         // then(검증): 어떠한 결과가 나와야 한다.
-        assertThat(refundAmount).isEqualTo(210000);
+        assertThat(refundAmount).isEqualTo(BigDecimal.valueOf(210000));
     }
 }
